# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Apex Fusion Data Provider API
  description: |
    Complete API specification for Apex Fusion blockchain data provider.
    
    This API must support three networks:
    - **Prime**: UTXO-based Cardano-like network
    - **Vector**: UTXO-based Cardano-like network  
    - **Nexus**: EVM-based Ethereum-like network
    
    All endpoints accept a `network` query parameter to specify which network to query.
    
    **Important Notes:**
    - All timestamps are in ISO 8601 format
    - All monetary values are in native currency units (ADA for UTXO, AP3X for EVM)
    - Hash values are hex strings (with or without 0x prefix)
    - Addresses can be in hex or bech32 format for UTXO networks
    - EVM addresses must include 0x prefix
  version: 1.0.0
  contact:
    name: Apex Fusion Explorer
    url: https://explorer.test
  license:
    name: Proprietary
servers:
  - url: https://api.apexfusion.org/v1
    description: Production API
  - url: https://api-staging.apexfusion.org/v1
    description: Staging API
  - url: http://localhost:4010
    description: Local development
...
paths:
...
  /transactions:
    get:
      tags:
        - Transactions
      summary: List transactions
      description: Get paginated list of transactions
      parameters:
        - $ref: '#/components/parameters/Network'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /transactions/latest:
    get:
      tags:
        - Transactions
      summary: Get latest transactions
      description: Get the most recent transactions (typically 10)
      parameters:
        - $ref: '#/components/parameters/Network'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /transactions/{hash}:
    get:
      tags:
        - Transactions
      summary: Get transaction details
      description: Get detailed information about a specific transaction
      parameters:
        - $ref: '#/components/parameters/Network'
        - name: hash
          in: path
          required: true
          description: Transaction hash (hex string)
          schema:
            type: string
            pattern: '^(0x)?[0-9a-fA-F]{64}$'
            example: "a1b2c3d4e5f6..."
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /tokens:
    get:
      tags:
        - Tokens
      summary: List tokens
      description: Get paginated list of tokens/assets
      parameters:
        - $ref: '#/components/parameters/Network'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokensResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tokens/{identifier}:
    get:
      tags:
        - Tokens
      summary: Get token details
      description: Get detailed information about a specific token. For UTXO networks, identifier is `policy_id.asset_name_hex`. For EVM networks, identifier is contract address.
      parameters:
        - $ref: '#/components/parameters/Network'
        - name: identifier
          in: path
          required: true
          description: Token identifier (policy_id.asset_name for UTXO, contract address for EVM)
          schema:
            type: string
            example: "a1b2c3d4e5f6..."
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tokens/{identifier}/holders:
    get:
      tags:
        - Tokens
      summary: Get token holders
      description: Get paginated list of token holders (EVM networks only)
      parameters:
        - $ref: '#/components/parameters/Network'
        - name: identifier
          in: path
          required: true
          description: Token contract address
          schema:
            type: string
            pattern: '^0x[0-9a-fA-F]{40}$'
            example: "0x1234567890abcdef..."
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenHoldersResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tokens/{identifier}/transfers:
    get:
      tags:
        - Tokens
      summary: Get token transfers
      description: Get paginated list of token transfers for a specific token (EVM networks only)
      parameters:
        - $ref: '#/components/parameters/Network'
        - name: identifier
          in: path
          required: true
          description: Token contract address
          schema:
            type: string
            pattern: '^0x[0-9a-fA-F]{40}$'
            example: "0x1234567890abcdef..."
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenTransfersResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /token-transfers:
    get:
      tags:
        - Tokens
      summary: List all token transfers
      description: Get paginated list of all token transfers (EVM networks only). Optionally filter by token address.
      parameters:
        - $ref: '#/components/parameters/Network'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: token
          in: query
          description: Filter by token contract address (optional)
          schema:
            type: string
            pattern: '^0x[0-9a-fA-F]{40}$'
            example: "0x1234567890abcdef..."
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenTransfersResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /epochs:
    get:
      tags:
        - Epochs
      summary: List epochs
      description: Get paginated list of epochs (UTXO networks only)
      parameters:
        - $ref: '#/components/parameters/Network'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpochsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /epochs/{number}:
    get:
      tags:
        - Epochs
      summary: Get epoch details
      description: Get detailed information about a specific epoch (UTXO networks only)
      parameters:
        - $ref: '#/components/parameters/Network'
        - name: number
          in: path
          required: true
          description: Epoch number
          schema:
            type: integer
            minimum: 0
            example: 42
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Epoch'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /pools:
    get:
      tags:
        - Pools
      summary: List staking pools
      description: Get paginated list of staking pools (UTXO networks only)
      parameters:
        - $ref: '#/components/parameters/Network'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: search
          in: query
          description: Search by pool ID, ticker, or name
          schema:
            type: string
            example: "POOL"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /pools/{id}:
    get:
      tags:
        - Pools
      summary: Get pool details
      description: Get detailed information about a specific staking pool (UTXO networks only)
      parameters:
        - $ref: '#/components/parameters/Network'
        - name: id
          in: path
          required: true
          description: Pool ID (hex string or bech32)
          schema:
            type: string
            example: "a1b2c3d4e5f6..."
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  pool:
                    $ref: '#/components/schemas/Pool'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /contracts:
    get:
      tags:
        - Contracts
      summary: List smart contracts
      description: Get paginated list of verified smart contracts (EVM networks only)
      parameters:
        - $ref: '#/components/parameters/Network'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: search
          in: query
          description: Search by contract name or address
          schema:
            type: string
            example: "MyContract"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /contracts/{address}:
    get:
      tags:
        - Contracts
      summary: Get contract details
      description: Get detailed information about a smart contract including ABI and source code (EVM networks only)
      parameters:
        - $ref: '#/components/parameters/Network'
        - name: address
          in: path
          required: true
          description: Contract address
          schema:
            type: string
            pattern: '^0x[0-9a-fA-F]{40}$'
            example: "0x1234567890abcdef..."
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    Network:
      name: network
      in: query
      required: true
      description: Network identifier
      schema:
        type: string
        enum: [prime, vector, nexus]
        example: "prime"
    Limit:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20
    Offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        minimum: 0
        default: 0
        example: 0
  schemas:
...
    Transaction:
      type: object
      required:
        - hash
        - block_height
        - status
      properties:
        hash:
          type: string
          description: Transaction hash (hex string)
          pattern: '^(0x)?[0-9a-fA-F]{64}$'
          example: "a1b2c3d4e5f6..."
        block_height:
          type: integer
          description: Block height containing this transaction
          example: 12345
        block_slot:
          type: integer
          nullable: true
          description: Block slot (UTXO networks only)
          example: 12345678
        slot:
          type: integer
          nullable: true
          description: Transaction slot (UTXO networks only, same as block_slot)
          example: 12345678
        epoch:
          type: integer
          nullable: true
          description: Epoch number (UTXO networks only)
          example: 42
        index:
          type: integer
          nullable: true
          description: Transaction index in block (for list endpoints)
          example: 0
        tx_index:
          type: integer
          nullable: true
          description: Transaction index in block (for detail endpoints)
          example: 0
        from_addresses:
          type: array
          items:
            type: string
          nullable: true
          description: Source addresses (hex format, for list endpoints)
          example: ["a1b2c3d4e5f6..."]
        to_addresses:
          type: array
          items:
            type: string
          nullable: true
          description: Destination addresses (hex format, for list endpoints)
          example: ["f6e5d4c3b2a1..."]
        from_addresses_bech32:
          type: array
          items:
            type: string
          nullable: true
          description: Source addresses (bech32 format, UTXO networks only, for list endpoints)
          example: ["addr1qx2fxv2umyhttkxyxp8x0dlpdt3k6cwng5pxj3jhsydzer3jcu5d8ps7zex2k2xt3uqxgjqnnj..."]
        to_addresses_bech32:
          type: array
          items:
            type: string
          nullable: true
          description: Destination addresses (bech32 format, UTXO networks only, for list endpoints)
          example: ["addr1qx2fxv2umyhttkxyxp8x0dlpdt3k6cwng5pxj3jhsydzer3jcu5d8ps7zex2k2xt3uqxgjqnnj..."]
        input_count:
          type: integer
          nullable: true
          description: Number of inputs (UTXO networks only, for list endpoints)
          example: 2
        output_count:
          type: integer
          nullable: true
          description: Number of outputs (UTXO networks only, for list endpoints)
          example: 1
        ada_amount:
          type: number
          nullable: true
          description: Transaction amount in native currency (for list endpoints)
          example: 100.5
        total_input_ada:
          type: number
          nullable: true
          description: Total input amount in native currency (for detail endpoints)
          example: 100.5
        total_output_ada:
          type: number
          nullable: true
          description: Total output amount in native currency (for detail endpoints)
          example: 100.33
        fee:
          type: number
          nullable: true
          description: Transaction fee in native currency (for list endpoints)
          example: 0.17
        fee_ada:
          type: number
          nullable: true
          description: Transaction fee in native currency (for detail endpoints)
          example: 0.17
        status:
          type: string
          enum: [success, failed, confirmed, invalid, pending]
          description: Transaction status
          example: "success"
        is_valid:
          type: boolean
          nullable: true
          description: Whether transaction is valid (UTXO networks only)
          example: true
        confirmations:
          type: integer
          nullable: true
          description: Number of confirmations
          example: 100
        timestamp:
          type: string
          format: date-time
          nullable: true
          description: Transaction timestamp
          example: "2024-01-15T10:30:00Z"
        block_hash:
          type: string
          nullable: true
          description: Block hash containing this transaction
          pattern: '^(0x)?[0-9a-fA-F]{64}$'
          example: "a1b2c3d4e5f6..."
        inputs:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/TransactionInput'
          description: Transaction inputs (for detail endpoints, UTXO networks only)
        outputs:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/TransactionOutput'
          description: Transaction outputs (for detail endpoints)
        gas_used:
          type: integer
          nullable: true
          description: Gas used (EVM networks only)
          example: 21000
        gas_price:
          type: number
          nullable: true
          description: Gas price in native currency (EVM networks only)
          example: 0.00000002
        gas_limit:
          type: integer
          nullable: true
          description: Gas limit (EVM networks only)
          example: 21000
        nonce:
          type: integer
          nullable: true
          description: Transaction nonce (EVM networks only)
          example: 5
        method:
          type: string
          nullable: true
          description: Transaction method signature (EVM networks only)
          example: "transfer"
        evmType:
          type: string
          nullable: true
          description: Transaction method signature (EVM networks only, alias for method)
          example: "transfer"
        error:
          type: string
          nullable: true
          description: Error message if transaction failed (EVM networks only)
          example: null
...
    TransactionsResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Token:
      type: object
      properties:
        address:
          type: string
          nullable: true
          description: Token contract address (EVM networks)
          example: "0x1234567890abcdef..."
        policy_id:
          type: string
          nullable: true
          description: Policy ID (UTXO networks)
          example: "a1b2c3d4e5f6..."
        asset_name:
          type: string
          nullable: true
          description: Asset name (UTXO networks)
          example: "TOKEN"
        asset_name_hex:
          type: string
          nullable: true
          description: Asset name in hex (UTXO networks)
          example: "544f4b454e"
        name:
          type: string
          example: "My Token"
        symbol:
          type: string
          example: "MTK"
        decimals:
          type: integer
          example: 18
        total_supply:
          type: string
          description: Total supply as string
          example: "1000000000000000000"
        total_supply_formatted:
          type: string
          description: Total supply formatted
          example: "1,000,000"
        quantity:
          type: string
          nullable: true
          description: Quantity (UTXO networks)
          example: "1000000"
        holders:
          type: integer
          nullable: true
          description: Number of holders
          example: 100
        type:
          type: string
          nullable: true
          description: Token type (EVM networks)
          example: "ERC-20"
        icon_url:
          type: string
          nullable: true
          description: Token icon URL
          example: "https://example.com/icon.png"
        metadata:
          type: object
          nullable: true
          description: Token metadata (UTXO networks)
        fingerprint:
          type: string
          nullable: true
          description: Token fingerprint (UTXO networks)
          example: "a1b2c3d4...e5f6g7h8"
    TokensResponse:
      type: object
      properties:
        tokens:
          type: array
          items:
            $ref: '#/components/schemas/Token'
        pagination:
          $ref: '#/components/schemas/Pagination'
    TokenHoldersResponse:
      type: object
      properties:
        holders:
          type: array
          items:
            type: object
            properties:
              address:
                type: string
                example: "0x1234567890abcdef..."
              value:
                type: string
                description: Balance as string
                example: "1000000000000000000"
              balance:
                type: string
                description: Balance as string (alias for value)
                example: "1000000000000000000"
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    TokenTransfersResponse:
      type: object
      properties:
        transfers:
          type: array
          items:
            $ref: '#/components/schemas/TokenTransfer'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    EpochsResponse:
      type: object
      properties:
        epochs:
          type: array
          items:
            type: object
            properties:
              epoch:
                type: integer
                description: Epoch number
                example: 42
              start_slot:
                type: integer
                description: Start slot number
                example: 12345678
              end_slot:
                type: integer
                description: End slot number
                example: 12345690
              start_time:
                type: string
                format: date-time
                nullable: true
                description: Epoch start time
                example: "2024-01-15T10:00:00Z"
              end_time:
                type: string
                format: date-time
                nullable: true
                description: Epoch end time
                example: "2024-01-15T11:00:00Z"
              blocks:
                type: integer
                nullable: true
                description: Number of blocks in epoch (alias for block_count)
                example: 1000
              block_count:
                type: integer
                nullable: true
                description: Number of blocks in epoch
                example: 1000
              transactions:
                type: integer
                nullable: true
                description: Number of transactions in epoch (alias for transaction_count)
                example: 5000
              transaction_count:
                type: integer
                nullable: true
                description: Number of transactions in epoch
                example: 5000
              fees:
                type: string
                nullable: true
                description: Estimated fees in epoch
                example: "850.00"
              is_current:
                type: boolean
                nullable: true
                description: Whether this is the current epoch
                example: false
        current_epoch:
          type: integer
          nullable: true
          description: Current epoch number
          example: 42
        pagination:
          $ref: '#/components/schemas/Pagination'
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid network parameter"
            details: "Network must be one of: prime, vector, nexus"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Block not found"
            details: "Block height 99999 does not exist"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            details: "Database connection failed"